+++ source_me.sh
@@ -1,0 +1,6 @@
+bin=$PWD/bin
+lib=$PWD/lib
+export PATH=$bin:$PATH
+echo 'adding' $bin 'to $PATH'
+export RUBYLIB=$lib
+echo 'adding' $lib 'to $RUBYLIB'