+++ test/start.rb
@@ -1,0 +1,3 @@
+#!/usr/bin/env ruby
+require 'collectd/interface/service'
+Collectd::Interface::Service.run!